// admin/src/types/admin.ts
export interface OrderItem {
  id: string;
  productName: string;
  variantSku: string;
  price: number;
  quantity: number;
  image: string;
}

export interface Order {
  id: string;
  orderNumber: string;
  customerName: string;
  customerEmail: string;
  date: string;
  totalAmount: number;
  status: 'PENDING' | 'PAID' | 'PROCESSING' | 'SHIPPED' | 'DELIVERED' | 'CANCELLED';
  paymentGateway: 'Razorpay' | 'Stripe' | 'COD';
  itemsCount: number;
  items: OrderItem[];
}